requirements.txt
alembic==1.0.10
backports.csv==1.0.7
chardet==3.0.4
Click==7.0
configparser==3.7.4
Django==2.2
docx-mailmerge==0.4.0
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.0
Flask-Uploads==0.2.1
future==0.17.1
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.3
Mako==1.0.9
Markdown==3.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
passlib==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-docx==0.8.10
python-editor==1.0.4
pytz==2019.1
PyYAML==5.1
six==1.12.0
SQLAlchemy==1.3.3
sqlparse==0.3.0
Werkzeug==0.15.2
WTForms==2.2.1
xlrd==1.2.0
